Fix "Phase Control" rendered twice
